chore: starting to refactor widgets for the pool page
